Shortest path in binary matrix
Given an n x n binary matrix grid, print the length of the shortest clear path in the matrix. If there is no clear path, print -1.

A clear path in a binary matrix is a path from the top-left cell (i.e., (0, 0)) to the bottom-right cell (i.e., (n - 1, n - 1)) such that:

All the visited cells of the path are 0.
All the adjacent cells of the path are 8-directionally connected (i.e., they are different and they share an edge or a corner).
The length of a clear path is the number of visited cells of this path.
Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give n.
The following n lines will give the binary matrix grid.
Constraints

1 <= T <= 1000
n == grid.length
n == grid[i].length
1 <= n <= 100
grid[i][j] is 0 or 1
Output Format

For each testcase, print the length of the shortest clear path in the matrix. If there is no clear path, print -1.

Sample Input 0

1
2
0 1
1 0
Sample Output 0

2
Explanation 0

image

Sample Input 1

1
3
0 0 0
1 1 0
1 1 0
Sample Output 1

4
Explanation 1

image

Sample Input 2

1
3
1 0 0
1 1 0
1 1 0
Sample Output 2

-1